requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cycler==0.10.0
discord.py==1.5.1
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.3
multidict==4.7.6
numpy==1.19.4
Pillow==8.0.1
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
typing-extensions==3.7.4.3
tzlocal==2.1
wincertstore==0.2
yarl==1.5.1
